% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_doc.R
\name{hello_2}
\alias{hello_2}
\title{#' Print a helpful message}
\usage{
hello_2()
}
\arguments{
\item{no}{params}
}
\value{
A helpful message
}
\description{
This is a trial function, but nonetheless has great value.
}
