% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackages.R
\name{mypak}
\alias{mypak}
\title{Install packages that don't exist; load them all}
\usage{
mypak(packages)
}
\arguments{
\item{packages}{A vector of packages}
}
\value{
all packages will be installed if necessary and possible, and loaded. Returns TRUE if successful for each package
}
\description{
Given a list of packages, installs the ones that aren't installed and loads them all.
}
\examples{
mypak(c("ggplot2","dplyr"))
}
