% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTextTree.R
\name{readTextTree}
\alias{readTextTree}
\title{Make a tree object with text}
\usage{
readTextTree(tree)
}
\arguments{
\item{tree}{A newick-formatted tree string}
}
\value{
ape tree object with your text as the tree
}
\description{
given a newick-formatted text string, returns an ape tree object of the shape
}
\examples{
textToTree("(((A,B),C),D);")
}
